pyproject.toml
[project]
name = "datagovgr-mcp-server"
version = "0.1.0"
description = "An MCP server for the data.gov.gr API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.6",
    "pydatagovgr>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.4",
]
